changes in image studio,Quotation excel format

+++ app/Http/Controllers/Panel/ProductAttributeController.php
@@ -6,12 +6,13 @@ use App\Http\Controllers\Controller;
 
 use Illuminate\Http\Request;
 use App\Models\ProductAttribute;
+use App\Models\ProductExtraInfo;
 use App\Models\ProductAttributeValue;
 use App\User;
 
 class ProductAttributeController extends Controller
 {
-    
+
 
     /**
      * Display a listing of the resource.
@@ -25,13 +26,13 @@ class ProductAttributeController extends Controller
              $length = $request->get('length');
          }
          $product_attributes = ProductAttribute::query();
-         
+
             if($request->get('search')){
                 $product_attributes->where('id','like','%'.$request->search.'%')
                 ->orWhere('name','like','%'.$request->search.'%')
                 ;
             }
-            
+
             if(AuthRole() != 'Admin') {
                 $product_attributes->where('user_id',auth()->id())->orWhere('user_id');
             }
@@ -49,17 +50,17 @@ class ProductAttributeController extends Controller
             $product_attributes = $product_attributes->paginate($length);
 
             if ($request->ajax()) {
-                return view('panel.product_attributes.load', ['product_attributes' => $product_attributes])->render();  
-                return view('panel.user_shop_items.includes.Properties', ['product_attributes' => $product_attributes])->render();  
+                return view('panel.product_attributes.load', ['product_attributes' => $product_attributes])->render();
+                return view('panel.user_shop_items.includes.Properties', ['product_attributes' => $product_attributes])->render();
             }
- 
+
         return view('panel.product_attributes.index', compact('product_attributes'));
     }
 
         public function print(Request $request){
             $product_attributes = collect($request->records['data']);
-                return view('panel.product_attributes.print', ['product_attributes' => $product_attributes])->render();  
-           
+                return view('panel.product_attributes.print', ['product_attributes' => $product_attributes])->render();
+
         }
 
     /**
@@ -71,7 +72,7 @@ class ProductAttributeController extends Controller
     {
         try{
             return view('panel.product_attributes.create');
-        }catch(\Exception $e){            
+        }catch(\Exception $e){
             return back()->with('error', 'There was an error: ' . $e->getMessage());
         }
     }
@@ -84,25 +85,35 @@ class ProductAttributeController extends Controller
      */
     public function store(Request $request)
     {
-        
+        // if ($request->has('request_from')) {
+        //     // $response = new Response();
+        //     return response()->json([
+        //         'status' => 'success',
+        //         'msg' => 'Product Property Created Successfully!'
+        //     ]);
+        // }
+
         // return magicstring($request->all());
+
         $this->validate($request, [
                         'name'     => 'required',
                         'type'     => 'sometimes',
                         'value'     => 'sometimes',
                     ]);
-        
+
         try{
+
+
             $UserRole = AuthRole();
             if ($UserRole != 'Admin') {
                 // `  Start Creating Custom Attribute....
                 $user = User::find(auth()->id());
                 $custom_arr_old = json_decode($user->custom_attriute_columns);
-                $custom_arr_new = [$request->name];
+                $custom_arr_new = [str_replace(' ','_',$request->name)];
 
                 // + creating Array for Adding in User Record
                 if ($custom_arr_old != null) {
-                    $upload_arr = array_merge((array) $custom_arr_old,$custom_arr_new); 
+                    $upload_arr = array_merge((array) $custom_arr_old,$custom_arr_new);
                 }else{
                     $upload_arr = $custom_arr_new;
                 }
@@ -113,34 +124,59 @@ class ProductAttributeController extends Controller
             }else{
                 $type = 0; // Admin Define Attribute
             }
- 
+
+
+            if (request()->has('type') && request()->get('type') == 'any_value') {
+                $value = 'any_value';
+            }elseif (request()->has('type') && request()->get('type') == 'uom') {
+                $value = 'uom';
+            }
+            else{
+                $value = null;
+            }
+
             $chkCount = ProductAttribute::where('name',$request->name)->where('user_id',$request->user_id)->get();
-            if (count($chkCount) == 0) {   
+            if (count($chkCount) == 0) {
                 // ! Uploading Attributes
                 $AttributValue = ProductAttribute::create([
-                    'name' => $request->get('name'),
+                    'name' => str_replace(' ','_',$request->get('name')),
                     'type' => $type,
-                    'value' => null,
+                    'value' => $value ?? NULL,
                     'user_id' => $request->get('user_id') ?? null,
                     'user_shop_id' => $request->get('user_shop_id') ?? null,
-                ]);   
-                
-                // - Checking Values
-                foreach (explode(",",$request->value[0]) as $key => $items) {
-                    ProductAttributeValue::create([
-                        'parent_id' => $AttributValue->id,
-                        'user_id' => $request->get('user_id') ?? null,
-                        'attribute_value' => $items,
-                    ]);
+                ]);
+
+                if ($value != 'any_value') {
+                    // - Checking Values
+                    foreach (explode(",",$request->value[0]) as $key => $items) {
+                        ProductAttributeValue::create([
+                            'parent_id' => $AttributValue->id,
+                            'user_id' => $request->get('user_id') ?? null,
+                            'attribute_value' => $items,
+                        ]);
+                    }
                 }
+
             }else{
                 return back()->with('error',"$request->name Property already exists in your Account.");
             }
-                
+
+
+        if ($request->has('request_from')) {
+            // $response = new Response();
+            return response()->json([
+                'status' => 'success',
+                'msg' => 'Product Property Created Successfully!'
+            ]);
+        }
+
+
             return redirect()->route('panel.product_attributes.index')->with('success','Product Attribute Created Successfully!');
-        }catch(\Exception $e){            
-            // return back()->with('error', 'There was an error: ' . $e->getMessage())->withInput($request->all());
-            echo $e->getMessage();
+        }catch(\Exception $e){
+            return back()->with('error', 'There was an error: ' . $e->getMessage())->withInput($request->all());
+
+
+            // throw $e;
         }
     }
 
@@ -154,7 +190,7 @@ class ProductAttributeController extends Controller
     {
         try{
             return view('panel.product_attributes.show',compact('product_attribute'));
-        }catch(\Exception $e){            
+        }catch(\Exception $e){
             return back()->with('error', 'There was an error: ' . $e->getMessage());
         }
     }
@@ -166,16 +202,16 @@ class ProductAttributeController extends Controller
      * @return  \Illuminate\Http\Response
      */
     public function edit(ProductAttribute $product_attribute)
-    {   
+    {
         try{
-            
+
             if (AuthRole() != 'Admin') {
                 if ($product_attribute->user_id != auth()->id() && $product_attribute->user_id != null) {
                     return back()->with('error',"Something went wrong.");
                 }
             }
             return view('panel.product_attributes.edit',compact('product_attribute'));
-        }catch(\Exception $e){            
+        }catch(\Exception $e){
             return back()->with('error', 'There was an error: ' . $e->getMessage());
         }
     }
@@ -189,18 +225,18 @@ class ProductAttributeController extends Controller
      */
     public function update(Request $request,ProductAttribute $product_attribute)
     {
-        
+
         // $this->validate($request, [
         //         '*' => 'required|alpha_num:ascii'
         //     ]);
-                
+
         try{
             $newrecord = 0;
-            $loopcount = 0; 
+            $loopcount = 0;
 
             //` Create New Values
             if (request()->has('newval') && request()->get('newval') != null) {
-                foreach (explode(",",request()->get('newval')) as $key => $items) {    
+                foreach (explode(",",request()->get('newval')) as $key => $items) {
                     $chk = ProductAttributeValue::where('attribute_value',$items)->where('parent_id',$product_attribute->id)->get();
                     if (count($chk) == 0) {
                         echo "New Value";
@@ -213,16 +249,16 @@ class ProductAttributeController extends Controller
                     }
                 }
             }
-            
+
             // Updating Existing Value
             foreach ($request->except(['_token','user_id','user_shop_id','name','newval']) as $key => $value) {
                 echo $value.newline();
                 ProductAttributeValue::where('id',$key)->update([
                     'attribute_value' => trim($value),
                 ]);
-                $loopcount++; 
+                $loopcount++;
             }
-            
+
 
             $product_attribute->visibility = $request->visibility ?? 0;
             $product_attribute->save();
@@ -231,7 +267,7 @@ class ProductAttributeController extends Controller
             $msg = "$newrecord are Created and $loopcount are Updated !!";
             // return back()->with('success',"Product Attribute Values Updated")->withInput($request->all());
             return back()->with('success',$msg)->withInput($request->all());
-        }catch(\Throwable $e){            
+        }catch(\Throwable $e){
             return back()->with('error', 'There was an error: ' . $e->getMessage())->withInput($request->all());
             // throw $e;
         }
@@ -242,10 +278,16 @@ class ProductAttributeController extends Controller
 
     public function deletevalue(Request $request,$product_attribute_value) {
 
+        $chk = ProductExtraInfo::where('attribute_value_id',$product_attribute_value)->get();
+
+        if ($chk->count() > 0) {
+            return back()->with('error','Delete property from associated products before removing the product property.');
+        }
+
         $user = auth()->user();
         $chk = ProductAttributeValue::where('user_id',$user->id)->whereId($product_attribute_value)->get();
         $name = '';
-        
+
         if (count($chk) == 0) {
             return back()->with('error',"Value doesn't exist, or isn't linked with your account!!");
         }else{
@@ -268,14 +310,20 @@ class ProductAttributeController extends Controller
     {
         try{
             if($product_attribute){
-                                      
+
+                $chk = ProductExtraInfo::where('attribute_value_id',$product_attribute->id)->get();
+
+                if ($chk->count() > 0) {
+                    return back()->with('error','Delete property from associated products before removing the product property.');
+                }
+
                 $product_attribute->delete();
                 ProductAttributeValue::where('parent_id',$product_attribute->id)->delete();
 
                 $user = User::find(auth()->id());
                 $arr = json_decode($user->custom_attriute_columns);
                 $index = array_search($product_attribute->name,$arr);
-                
+
                 unset($arr[$index]);
                 $user->custom_attriute_columns = json_encode($arr);
                 $user->save();
